+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG
 * @heff added and `extends` function for external subclassing ([view](https://github.com/videojs/video.js/pull/2078))
 * @forbesjo added the `scrubbing` property ([view](https://github.com/videojs/video.js/pull/2080))
 * @heff switched to border-box sizing for all player elements ([view](https://github.com/videojs/video.js/pull/2082))
+* @forbesjo added a vjs-button class to button controls ([view](https://github.com/videojs/video.js/pull/2084))
 
 --------------------